refactor: migrate functions out of server module
